Move types back in, but keep docs
